Feat: Uploading script
Script requests API response data from Google Maps API and extracts the latitude and longitude coordinates of the location input (e.g. San Francisco). The coordinates are used as arguments for the Foursquare API to get the closest restaurant relating to the mealtype argument. It finds the first picture of that restaurant, or if unavailable, returns a picture of a cheeseburger.